refactor into reader writer storage
