+++ rosbag2_py/package.xml
@@ -14,6 +14,7 @@
   <depend>python_cmake_module</depend>
   <depend>rosbag2_compression</depend>
   <depend>rosbag2_cpp</depend>
+  <depend>rosbag2_storage</depend>
 
   <test_depend>ament_lint_auto</test_depend>
   <test_depend>ament_lint_common</test_depend>
