+++ rosbag2_py/src/rosbag2_py/_reader.cpp
@@ -23,14 +23,10 @@
 #include <vector>
 
 #include "rosbag2_compression/sequential_compression_reader.hpp"
-#include "rosbag2_compression/sequential_compression_writer.hpp"
 #include "rosbag2_cpp/converter_options.hpp"
 #include "rosbag2_cpp/readers/sequential_reader.hpp"
 #include "rosbag2_cpp/reader.hpp"
 #include "rosbag2_cpp/storage_options.hpp"
-#include "rosbag2_cpp/writer.hpp"
-#include "rosbag2_cpp/writers/sequential_writer.hpp"
-#include "rosbag2_storage/ros_helper.hpp"
 #include "rosbag2_storage/storage_filter.hpp"
 #include "rosbag2_storage/topic_metadata.hpp"
 
@@ -89,57 +85,10 @@ public:
 protected:
   std::unique_ptr<rosbag2_cpp::Reader> reader_;
 };
-
-template<typename T>
-class Writer
-{
-public:
-  Writer()
-  : writer_(std::make_unique<rosbag2_cpp::Writer>(std::make_unique<T>()))
-  {
-  }
-
-  void open(
-    rosbag2_cpp::StorageOptions & storage_options,
-    rosbag2_cpp::ConverterOptions & converter_options)
-  {
-    writer_->open(storage_options, converter_options);
-  }
-
-  void create_topic(const rosbag2_storage::TopicMetadata & topic_with_type)
-  {
-    writer_->create_topic(topic_with_type);
-  }
-
-  void remove_topic(const rosbag2_storage::TopicMetadata & topic_with_type)
-  {
-    writer_->remove_topic(topic_with_type);
-  }
-
-  /// Write a serialized message to a bag file
-  void write(
-    const std::string & topic_name, const std::string & message,
-    const rcutils_time_point_value_t & time_stamp)
-  {
-    auto bag_message =
-      std::make_shared<rosbag2_storage::SerializedBagMessage>();
-
-    bag_message->topic_name = topic_name;
-    bag_message->serialized_data =
-      rosbag2_storage::make_serialized_message(message.c_str(), message.length());
-    bag_message->time_stamp = time_stamp;
-
-    writer_->write(bag_message);
-  }
-
-protected:
-  std::unique_ptr<rosbag2_cpp::Writer> writer_;
-};
-
 }  // namespace rosbag2_py
 
-PYBIND11_MODULE(_rosbag2_py, m) {
-  m.doc() = "Python wrapper of the rosbag2_cpp API";
+PYBIND11_MODULE(_reader, m) {
+  m.doc() = "Python wrapper of the rosbag2_cpp reader API";
 
   pybind11::class_<rosbag2_py::Reader<rosbag2_cpp::readers::SequentialReader>>(
     m, "SequentialReader")
@@ -171,54 +120,4 @@ PYBIND11_MODULE(_rosbag2_py, m) {
   .def(
     "reset_filter",
     &rosbag2_py::Reader<rosbag2_compression::SequentialCompressionReader>::reset_filter);
-
-  pybind11::class_<rosbag2_py::Writer<rosbag2_cpp::writers::SequentialWriter>>(
-    m, "SequentialWriter")
-  .def(pybind11::init())
-  .def("open", &rosbag2_py::Writer<rosbag2_cpp::writers::SequentialWriter>::open)
-  .def("write", &rosbag2_py::Writer<rosbag2_cpp::writers::SequentialWriter>::write)
-  .def("remove_topic", &rosbag2_py::Writer<rosbag2_cpp::writers::SequentialWriter>::remove_topic)
-  .def("create_topic", &rosbag2_py::Writer<rosbag2_cpp::writers::SequentialWriter>::create_topic);
-
-  pybind11::class_<rosbag2_py::Writer<rosbag2_compression::SequentialCompressionWriter>>(
-    m, "SequentialCompressionWriter")
-  .def(pybind11::init())
-  .def("open", &rosbag2_py::Writer<rosbag2_compression::SequentialCompressionWriter>::open)
-  .def("write", &rosbag2_py::Writer<rosbag2_compression::SequentialCompressionWriter>::write)
-  .def(
-    "remove_topic",
-    &rosbag2_py::Writer<rosbag2_compression::SequentialCompressionWriter>::remove_topic)
-  .def(
-    "create_topic",
-    &rosbag2_py::Writer<rosbag2_compression::SequentialCompressionWriter>::create_topic);
-
-  pybind11::class_<rosbag2_cpp::ConverterOptions>(m, "ConverterOptions")
-  .def(pybind11::init())
-  .def_readwrite(
-    "input_serialization_format",
-    &rosbag2_cpp::ConverterOptions::input_serialization_format)
-  .def_readwrite(
-    "output_serialization_format",
-    &rosbag2_cpp::ConverterOptions::output_serialization_format);
-
-  pybind11::class_<rosbag2_cpp::StorageOptions>(m, "StorageOptions")
-  .def(pybind11::init())
-  .def_readwrite("uri", &rosbag2_cpp::StorageOptions::uri)
-  .def_readwrite("storage_id", &rosbag2_cpp::StorageOptions::storage_id)
-  .def_readwrite(
-    "max_bagfile_size",
-    &rosbag2_cpp::StorageOptions::max_bagfile_size);
-
-  pybind11::class_<rosbag2_storage::StorageFilter>(m, "StorageFilter")
-  .def(pybind11::init())
-  .def_readwrite("topics", &rosbag2_storage::StorageFilter::topics);
-
-  pybind11::class_<rosbag2_storage::TopicMetadata>(m, "TopicMetadata")
-  .def(pybind11::init())
-  .def_readwrite("name", &rosbag2_storage::TopicMetadata::name)
-  .def_readwrite("type", &rosbag2_storage::TopicMetadata::type)
-  .def_readwrite(
-    "serialization_format",
-    &rosbag2_storage::TopicMetadata::serialization_format)
-  .def("equals", &rosbag2_storage::TopicMetadata::operator==);
 }
