+++ rosbag2_py/src/rosbag2_py/_storage.cpp
@@ -1,0 +1,62 @@
+// Copyright 2020 Open Source Robotics Foundation, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#include <pybind11/pybind11.h>
+#include <pybind11/stl.h>
+
+#include <chrono>
+#include <map>
+#include <memory>
+#include <string>
+#include <utility>
+#include <vector>
+
+#include "rosbag2_cpp/converter_options.hpp"
+#include "rosbag2_cpp/storage_options.hpp"
+#include "rosbag2_storage/storage_filter.hpp"
+#include "rosbag2_storage/topic_metadata.hpp"
+
+PYBIND11_MODULE(_storage, m) {
+  m.doc() = "Python wrapper of the rosbag2 utilities API";
+
+  pybind11::class_<rosbag2_cpp::ConverterOptions>(m, "ConverterOptions")
+  .def(pybind11::init())
+  .def_readwrite(
+    "input_serialization_format",
+    &rosbag2_cpp::ConverterOptions::input_serialization_format)
+  .def_readwrite(
+    "output_serialization_format",
+    &rosbag2_cpp::ConverterOptions::output_serialization_format);
+
+  pybind11::class_<rosbag2_cpp::StorageOptions>(m, "StorageOptions")
+  .def(pybind11::init())
+  .def_readwrite("uri", &rosbag2_cpp::StorageOptions::uri)
+  .def_readwrite("storage_id", &rosbag2_cpp::StorageOptions::storage_id)
+  .def_readwrite(
+    "max_bagfile_size",
+    &rosbag2_cpp::StorageOptions::max_bagfile_size);
+
+  pybind11::class_<rosbag2_storage::StorageFilter>(m, "StorageFilter")
+  .def(pybind11::init())
+  .def_readwrite("topics", &rosbag2_storage::StorageFilter::topics);
+
+  pybind11::class_<rosbag2_storage::TopicMetadata>(m, "TopicMetadata")
+  .def(pybind11::init())
+  .def_readwrite("name", &rosbag2_storage::TopicMetadata::name)
+  .def_readwrite("type", &rosbag2_storage::TopicMetadata::type)
+  .def_readwrite(
+    "serialization_format",
+    &rosbag2_storage::TopicMetadata::serialization_format)
+  .def("equals", &rosbag2_storage::TopicMetadata::operator==);
+}
